refactor: Transaction Service 추상화
